backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary>=2.9.10
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
pydantic>=2.5.0
pydantic-settings==2.1.0
email-validator>=2.0.0
websockets>=10.4

